feat: déplacement tests dans un fichier séparé
