+++ tests/Functional/EquipmentControllerTest.php
@@ -964,171 +964,4 @@ final class EquipmentControllerTest extends AbstractWebTestCase
         // La fédération est prioritaire : création réussie avec redirect
         $this->assertResponseRedirects('/equipment');
     }
-
-    // -----------------------------------------------------------------------
-    // Boutons "Retour" / "Annuler" — test du Referer via la session
-    // -----------------------------------------------------------------------
-
-    public function testShowPageBackButtonAlwaysLinksToEquipmentList(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-        $crawler = $this->client->request(Request::METHOD_GET, '/equipment/'.$this->gloveAId);
-        $this->assertResponseIsSuccessful();
-
-        $backLink = $crawler->selectLink('Retour à la liste');
-        $this->assertSame('/equipment', $backLink->attr('href'));
-        $this->assertNull($backLink->attr('data-history-back'), 'The back button should NOT use data-history-back');
-    }
-
-    public function testEditPageFromShowPageBackButtonLinksToShowPage(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-        $editUrl = '/equipment/'.$this->gloveAId.'/edit';
-        $showUrl = '/equipment/'.$this->gloveAId;
-
-        $crawler = $this->client->request(Request::METHOD_GET, $editUrl, [], [], [
-            'HTTP_REFERER' => $showUrl,
-        ]);
-        $this->assertResponseIsSuccessful();
-
-        $this->assertSame($showUrl, $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame($showUrl, $crawler->selectLink('Annuler')->attr('href'));
-    }
-
-    public function testEditPageFromListBackButtonLinksToList(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-        $editUrl = '/equipment/'.$this->gloveAId.'/edit';
-
-        $crawler = $this->client->request(Request::METHOD_GET, $editUrl, [], [], [
-            'HTTP_REFERER' => '/equipment',
-        ]);
-        $this->assertResponseIsSuccessful();
-
-        $this->assertSame('/equipment', $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame('/equipment', $crawler->selectLink('Annuler')->attr('href'));
-    }
-
-    public function testEditPageUnknownRefererDefaultsToList(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-        $editUrl = '/equipment/'.$this->gloveAId.'/edit';
-
-        $crawler = $this->client->request(Request::METHOD_GET, $editUrl, [], [], [
-            'HTTP_REFERER' => 'https://evil.com/phishing',
-        ]);
-        $this->assertResponseIsSuccessful();
-
-        $this->assertSame('/equipment', $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame('/equipment', $crawler->selectLink('Annuler')->attr('href'));
-    }
-
-    public function testEditPageNoRefererDefaultsToList(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-        $editUrl = '/equipment/'.$this->gloveAId.'/edit';
-
-        $crawler = $this->client->request(Request::METHOD_GET, $editUrl);
-        $this->assertResponseIsSuccessful();
-
-        $this->assertSame('/equipment', $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame('/equipment', $crawler->selectLink('Annuler')->attr('href'));
-    }
-
-    public function testEditFormPostWithErrorsPreservesBackHrefFromSession(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-        $editUrl = '/equipment/'.$this->gloveAId.'/edit';
-        $showUrl = '/equipment/'.$this->gloveAId;
-
-        // Step 1: open edit from show page → backHref=show is stored in session
-        $crawler = $this->client->request(Request::METHOD_GET, $editUrl, [], [], [
-            'HTTP_REFERER' => $showUrl,
-        ]);
-        $this->assertResponseIsSuccessful();
-        $this->assertSame($showUrl, $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame($showUrl, $crawler->selectLink('Annuler')->attr('href'));
-
-        // Step 2: submit form with all owners empty → triggers ExactlyOneOwner error
-        // The Referer is now the edit page itself (or empty), but session should keep show
-        $crawler = $this->client->request(Request::METHOD_POST, $editUrl, [
-            'equipment_form' => [
-                'ownerFederation' => '',
-                'ownerRegion'     => '',
-                'ownerClub'       => '',
-                'borrowerClub'    => '',
-                'borrowerMember'  => '',
-                'state'           => 'used',
-                'notes'           => '',
-                'glove_form'      => ['nb_fingers' => '3', 'size' => '7'],
-            ],
-        ]);
-
-        $content = (string) $this->client->getResponse()->getContent();
-        $this->assertStringContainsString('Vous devez choisir un propriétaire', $content);
-        $this->assertSame($showUrl, $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame($showUrl, $crawler->selectLink('Annuler')->attr('href'));
-    }
-
-    public function testDifferentEquipmentsHaveIndependentBackHref(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-
-        $showA = '/equipment/'.$this->gloveAId;
-        $editA = '/equipment/'.$this->gloveAId.'/edit';
-        $showB = '/equipment/'.$this->gloveBId;
-        $editB = '/equipment/'.$this->gloveBId.'/edit';
-
-        // Step 1: open edit for gloveA from show page → stores backHref=showA under equipment_edit_back_href_{gloveAId}
-        $crawler = $this->client->request(Request::METHOD_GET, $editA, [], [], [
-            'HTTP_REFERER' => $showA,
-        ]);
-        $this->assertSame($showA, $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame($showA, $crawler->selectLink('Annuler')->attr('href'));
-
-        // Step 2: open edit for gloveB from show page → stores backHref=showB under equipment_edit_back_href_{gloveBId}
-        $crawler = $this->client->request(Request::METHOD_GET, $editB, [], [], [
-            'HTTP_REFERER' => $showB,
-        ]);
-        $this->assertSame($showB, $crawler->selectLink('Retour')->attr('href'));
-        $this->assertSame($showB, $crawler->selectLink('Annuler')->attr('href'));
-
-        // Step 3: POST on gloveA edit with errors → should read session key for gloveA, NOT gloveB
-        $crawler = $this->client->request(Request::METHOD_POST, $editA, [
-            'equipment_form' => [
-                'ownerFederation' => '',
-                'ownerRegion'     => '',
-                'ownerClub'       => '',
-                'borrowerClub'    => '',
-                'borrowerMember'  => '',
-                'state'           => 'used',
-                'notes'           => '',
-                'glove_form'      => ['nb_fingers' => '3', 'size' => '7'],
-            ],
-        ]);
-        $this->assertStringContainsString(
-            'Vous devez choisir un propriétaire',
-            (string) $this->client->getResponse()->getContent()
-        );
-        $this->assertSame(
-            $showA,
-            $crawler->selectLink('Retour')->attr('href'),
-            'Back button should link to gloveA show page, not gloveB'
-        );
-        $this->assertSame(
-            $showA,
-            $crawler->selectLink('Annuler')->attr('href'),
-            'Back button should link to gloveA show page, not gloveB'
-        );
-    }
-
-    public function testCreatePageBackButtonLinksToList(): void
-    {
-        $this->loginAs(AppFixtures::USER_ADMIN);
-        $crawler = $this->client->request(Request::METHOD_GET, '/equipment/create');
-        $this->assertResponseIsSuccessful();
-
-        $this->assertSame('/equipment', $crawler->selectLink('Retour à la liste')->attr('href'));
-        $this->assertSame('/equipment', $crawler->selectLink('Annuler')->attr('href'));
-    }
 }
